[CHANGE] seperated material and model resource pools into seperate files [CHANGE] seperated material and model resource pool backends into seperate files [FIX] Material pool now returns pointers to handles [FIX] Constant buffer pool now deletes handle after deallocating it [FIX] Command queue destroy function now checks if the queue isn't a nullptr before deleting it
