+++ src/scene_graph/scene_graph.hpp
@@ -6,7 +6,7 @@
 #include <DirectXMath.h>
 
 #include "../util/defines.hpp"
-#include "../resource_pool.hpp"
+#include "../resource_pool_model.hpp"
 #include "../resource_pool_constant_buffer.hpp"
 
 namespace wr
